# Optimize udivmod_1e19 function