@dtolnay 2021-08-22
Published in itoa 0.4.8.